Making it serve /out if on heroku else ../out
This is a temp solution, while I see if it actually works on Heroku.
I am thinking create a env var with base directory in it as a
long term solution, but this will do for now.

This is all assuming the fails on Heroku are to do with incorrect file
paths.